Buttons

+++ app/src/main/java/com/pawlowski/imucollector/ui/MainState.kt
@@ -7,7 +7,7 @@ data class MainState(
     val isTrackingInProgress: Boolean = false,
     val isSending: Boolean = false,
     val lastStartTime: Long? = null,
-    val lastPrediction: Map<ActivityType, Float>? = null,
+    val lastPrediction: Map<ActivityType, Int>? = null,
     val runMode: RunMode = RunMode.TRAINING,
     val error: String? = null,
 )
